requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
Pillow==10.1.0
python-docx==0.8.11
reportlab==4.0.7
aiofiles==23.2.0
jinja2==3.1.2
python-magic==0.4.27
pandas==2.1.4
